refactor(Item): DamageDurability 함수의 파라미터 타입 변경
내구도 감소량의 데이터 타입을 `int32`에서 `float`으로 수정하여 유연성 개선.
- `TSInteractionActorBase::DamageDurability` 함수의 파라미터 타입 변경
- 관련 헤더 파일 선언 부분 동기화
- 내구도 계산 시 강제 형변환 `(int)` 추가

+++ Source/TinySurvivor/Private/Item/TSInteractionActorBase.cpp
@@ -124,14 +124,14 @@ void ATSInteractionActorBase::InitializeFromBuildingData(const FBuildingData& Bu
 	}
 }
 
-void ATSInteractionActorBase::DamageDurability(UAbilitySystemComponent* ASC, int32 DamageAmount)
+void ATSInteractionActorBase::DamageDurability(UAbilitySystemComponent* ASC, float DamageAmount)
 {
 	if (!HasAuthority())
 	{
 		return;
 	}
 	// 내구도 업데이트
-	ItemInstance.CurrentDurability -= DamageAmount;
+	ItemInstance.CurrentDurability -= (int)DamageAmount;
 
 	// 플레이어면 공격한 아이템 내구도 업데이트
 	if (ASC->GetAvatarActor()->IsA(ATSCharacter::StaticClass()))
